move fixtures to conftest. Remove project uuid from list all output
